+++ packages/webview_flutter/webview_flutter_android/android/src/test/java/io/flutter/plugins/webviewflutter/WebChromeClientTest.java
@@ -120,9 +120,13 @@ public class WebChromeClientTest {
   public void onPermissionRequest() {
     final PermissionRequest mockRequest = mock(PermissionRequest.class);
     instanceManager.addDartCreatedInstance(mockRequest, 10);
-
     webChromeClient.onPermissionRequest(mockRequest);
-
     verify(mockFlutterApi).onPermissionRequest(eq(webChromeClient), eq(mockRequest), any());
+  }
+
+  @Test
+  public void onGeolocationPermissionsHidePrompt() {
+    webChromeClient.onGeolocationPermissionsHidePrompt();
+    verify(mockFlutterApi).onGeolocationPermissionsHidePrompt(eq(webChromeClient), any());
   }
 }