+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,6 +1,11 @@
-## 3.6.0
+## 3.7.0
 
 * Adds support for `PlatformWebViewController.setOnPlatformPermissionRequest`.
+* 
+## 3.6.0
+
+* Adds support for handling geolocation permissions. See
+  `AndroidWebViewController.setOnGeolocationPermissionsShowPrompt`.
 
 ## 3.5.3